Use a component to start the listener, stop listening on unmount.
