+++ src/main.tsx
@@ -21,6 +21,7 @@ import PasswordHandler from './components/PasswordHandler'
 import RecoveryKeyHandler from './components/RecoveryKeyHandler'
 import SpendingAuthorizationHandler from './components/SpendingAuthorizationHandler'
 import AuthRedirector from './navigation/AuthRedirector'
+import { WalletBridge } from './WalletBridge'
 
 // Define queries for responsive design
 const queries = {
@@ -39,6 +40,7 @@ if (rootElement) {
     <React.StrictMode>
       <UserContextProvider>
         <WalletContextProvider>
+          <WalletBridge />
           <ExchangeRateContextProvider>
             <Router>
               <AuthRedirector />
